Changed: Consolidate agent types under a shared module and improve docs readability
Reorganize agent configuration and error types into a top-level types module and rewrite module docs to be clearer, shorter, and easier to scan.

Changes:
- Moved agent config and load error definitions into `src/types/config.rs` and `src/types/error.rs`
- Added `src/types/mod.rs` to centralize re-exports while preserving the public API
- Updated internal imports across loader, parser, catalog, and extensions to use the new module layout
- Rewrote module-level documentation with headings, bullets, and an accessible YAML frontmatter example

Benefits:
- Makes type ownership and module structure easier to navigate and maintain
- Reduces confusion from outdated descriptions by documenting current parsing and naming behavior
