+++ src/llm-coding-tools-agents/src/lib.rs
@@ -1,16 +1,13 @@
 #![doc = include_str!(concat!("../", env!("CARGO_PKG_README")))]
 
 mod catalog;
-mod config;
-mod error;
 mod extensions;
 mod loader;
 mod parser;
+mod types;
 
 pub use catalog::AgentCatalog;
-pub use config::{AgentConfig, AgentMode, PermissionRule};
-pub use error::AgentLoadError;
-pub use error::AgentLoadResult;
 pub use extensions::RulesetExt;
 pub use loader::AgentLoader;
 pub use parser::AgentParseError;
+pub use types::{AgentConfig, AgentLoadError, AgentLoadResult, AgentMode, PermissionRule};
